Update 3dwakecode.m
We now iterate to create k values from which a frequency is calculated rather than the other way around. This seems to solve some of the issues that we had, but the shallow water case is still being weird. There is also code for 4 moving point sources that I am using to model a box. The contributions from the "middle" of the box is unphysical. Also don't let starting p value ,from which a k values is generated, be too small, as the phase velocity blows up when k near 0.